move users integration tests to separated suite
